Feature/categories

+++ db/migrate/20220524153357_create_operations.rb
@@ -2,7 +2,7 @@ class CreateOperations < ActiveRecord::Migration[7.0]
   def change
     create_table :operations do |t|
       t.string :name
-      t.integer :amount, :default => 0
+      t.float :amount, :default => 0.00
 
       t.timestamps
     end
